+++ src/Exception/TemplateNotWritten.php
@@ -4,6 +4,6 @@ declare(strict_types=1);
 
 namespace Madapaja\TwigModule\Exception;
 
-class TemplateAlreadyLoaded extends RuntimeException
+class TemplateNotWritten extends RuntimeException
 {
 }